+++ lib/screens/signals_calendar_day.dart
@@ -5,13 +5,11 @@ import 'package:periodik/utils/date_time.dart';
 class SignalsCalendarDay extends StatelessWidget {
   const SignalsCalendarDay({
     required this.date,
-    required this.focusedDate,
     required this.points,
     super.key,
   });
 
   final DateTime date;
-  final DateTime focusedDate;
   final List<PointWithSignal> points;
 
   @override
@@ -19,7 +17,6 @@ class SignalsCalendarDay extends StatelessWidget {
     final today = DateTime.now();
     final theme = Theme.of(context);
     final isToday = today.isSameDayAs(date);
-    final isOtherMonth = focusedDate.month != date.month;
     return DecoratedBox(
       position: DecorationPosition.foreground,
       decoration: BoxDecoration(
@@ -31,32 +28,31 @@ class SignalsCalendarDay extends StatelessWidget {
               )
             : null,
       ),
-      child: Opacity(
-        opacity: isOtherMonth ? 0.6 : 1,
-        child: Center(
-          child: ListView.builder(
-            itemCount: points.length + 1,
-            itemBuilder: (context, index) {
-              if (index == 0) {
-                return Padding(
-                  padding: const EdgeInsets.all(2.0),
-                  child: Text(date.day.toString()),
-                );
-              }
-              final theme = Theme.of(context);
-              final point = points[index - 1];
-              return ColoredBox(
-                color: point.point.state
-                    ? theme.colorScheme.errorContainer
-                    : theme.colorScheme.secondaryContainer,
-                child: Text(point.signal.name,
-                    style: TextStyle(
-                        color: point.point.state
-                            ? theme.colorScheme.onErrorContainer
-                            : theme.colorScheme.onSecondaryContainer)),
+      child: Center(
+        child: ListView.builder(
+          primary: false,
+          // shrinkWrap: true,
+          itemCount: points.length + 1,
+          itemBuilder: (context, index) {
+            if (index == 0) {
+              return Padding(
+                padding: const EdgeInsets.all(2.0),
+                child: Text(date.day.toString()),
               );
-            },
-          ),
+            }
+            final theme = Theme.of(context);
+            final point = points[index - 1];
+            return ColoredBox(
+              color: point.point.state
+                  ? theme.colorScheme.errorContainer
+                  : theme.colorScheme.secondaryContainer,
+              child: Text(point.signal.name,
+                  style: TextStyle(
+                      color: point.point.state
+                          ? theme.colorScheme.onErrorContainer
+                          : theme.colorScheme.onSecondaryContainer)),
+            );
+          },
         ),
       ),
     );