+++ lib/widgets/calendar/calendar_day.dart
@@ -10,7 +10,6 @@ enum CalendarDayState {
 class CalendarDay extends StatelessWidget {
   const CalendarDay({
     required this.date,
-    required this.focusedDate,
     required this.onPressed,
     this.onLongPressed,
     this.state = CalendarDayState.none,
@@ -18,8 +17,6 @@ class CalendarDay extends StatelessWidget {
   });
 
   final DateTime date;
-
-  final DateTime focusedDate;
 
   final VoidCallback onPressed;
 
@@ -31,7 +28,6 @@ class CalendarDay extends StatelessWidget {
   Widget build(BuildContext context) {
     final today = DateTime.now();
     final isToday = today.isSameDayAs(date);
-    final isOtherMonth = focusedDate.month != date.month;
     final theme = Theme.of(context);
 
     final Color? backgroundColor;
@@ -50,28 +46,25 @@ class CalendarDay extends StatelessWidget {
       onTap: onPressed,
       onLongPress: onLongPressed,
       borderRadius: BorderRadius.circular(4),
-      child: Opacity(
-        opacity: isOtherMonth ? 0.6 : 1,
-        child: DecoratedBox(
-          decoration: BoxDecoration(
-            color: backgroundColor,
-            borderRadius: BorderRadius.circular(4),
-            border: isToday
-                ? Border.all(
-                    color: theme.colorScheme.primary,
-                    width: 2,
-                  )
-                : null,
-          ),
-          child: SizedBox.expand(
-            child: Center(
-              child: Text(
-                date.day.toString(),
-                style: isToday
-                    ? theme.textTheme.titleLarge!
-                        .copyWith(color: theme.colorScheme.primary)
-                    : null,
-              ),
+      child: DecoratedBox(
+        decoration: BoxDecoration(
+          color: backgroundColor,
+          borderRadius: BorderRadius.circular(4),
+          border: isToday
+              ? Border.all(
+                  color: theme.colorScheme.primary,
+                  width: 2,
+                )
+              : null,
+        ),
+        child: SizedBox.expand(
+          child: Center(
+            child: Text(
+              date.day.toString(),
+              style: isToday
+                  ? theme.textTheme.titleLarge!
+                      .copyWith(color: theme.colorScheme.primary)
+                  : null,
             ),
           ),
         ),